ISSUE #5544 - rename context
